+++ nengo_spa/tests/test_actions.py
@@ -33,13 +33,13 @@ def test_new_action_syntax(Simulator, seed, plt, rng):
         nengo.Connection(node1, model.buff1.input)
         nengo.Connection(node2, model.buff2.input)
 
-        actions = spa.Actions(
+        spa.Actions((
             'ctrl = input',
             'state = buff1',
             'dot(ctrl, A) --> buff3=buff1',
             'dot(ctrl, B) --> buff3=buff2',
-            'dot(ctrl, C) --> buff3=buff1*buff2')
-        actions.build()
+            'dot(ctrl, C) --> buff3=buff1*buff2'
+        ))
 
         state_probe = nengo.Probe(model.state.output, synapse=0.03)
         buff3_probe = nengo.Probe(model.buff3.output, synapse=0.03)
@@ -83,11 +83,11 @@ def test_dot_product(Simulator, seed, plt):
 
         model.stimulus = spa.Encode(lambda t: 'A' if t <= 0.3 else 'B', vocab=d)
 
-        actions = spa.Actions(
+        spa.Actions((
             'state_a = A',
             'state_b = stimulus',
-            'result = dot(state_a, state_b)')
-        actions.build()
+            'result = dot(state_a, state_b)'
+        ))
 
         p = nengo.Probe(model.result.output, synapse=0.03)
 
@@ -113,59 +113,49 @@ class TestExceptions():
 
     def test_invalid_types_in_binary_operation(self, model):
         with model:
-            actions = spa.Actions('state_c = state_a + state_b')
             with pytest.raises(SpaTypeError):
-                actions.build()
+                spa.Actions(('state_c = state_a + state_b',))
 
     def test_approx_inverse_of_scalar(self, model):
         with model:
-            actions = spa.Actions('state_c = ~scalar')
             with pytest.raises(SpaTypeError):
-                actions.build()
+                spa.Actions(('state_c = ~scalar',))
 
     def test_dot_product_incompatiple_vocabs(self, model):
         with model:
-            actions = spa.Actions('scalar = dot(state_a, state_b)')
             with pytest.raises(SpaTypeError):
-                actions.build()
+                spa.Actions(('scalar = dot(state_a, state_b)',))
 
     def test_dot_product_first_argument_invalid(self, model):
         with model:
-            actions = spa.Actions('scalar = dot(scalar, state_b)')
             with pytest.raises(SpaTypeError):
-                actions.build()
+                spa.Actions(('scalar = dot(scalar, state_b)',))
 
     def test_dot_product_second_argument_invalid(self, model):
         with model:
-            actions = spa.Actions('scalar = dot(state_a, scalar)')
             with pytest.raises(SpaTypeError):
-                actions.build()
+                spa.Actions(('scalar = dot(state_a, scalar)',))
 
     @pytest.mark.parametrize('method', ['reinterpret', 'translate'])
     def test_cast_type_inference_not_possible(self, model, method):
         with model:
-            actions = spa.Actions('scalar = dot({}(state_a), A)'.format(
-                method))
             with pytest.raises(SpaTypeError):
-                actions.build()
+                spa.Actions(('scalar = dot({}(state_a), A)'.format(method),))
 
     @pytest.mark.parametrize('method', ['reinterpret', 'translate'])
     def test_cast_scalar(self, model, method):
         with model:
-            actions = spa.Actions('state_a = {}(scalar)'.format(
-                method))
             with pytest.raises(SpaTypeError):
-                actions.build()
+                spa.Actions(('state_a = {}(scalar)'.format(method),))
 
     def test_reinterpret_non_matching_dimensions(self, model):
         with model:
-            actions = spa.Actions('state_a = reinterpret(state_b)')
             with pytest.raises(SpaTypeError):
-                actions.build()
+                spa.Actions(('state_a = reinterpret(state_b)',))
 
 
 def test_access_actions():
-    actions = spa.Actions('a = b', 'b = c', named='c = d')
+    actions = spa.Actions(('a = b', 'b = c'), {'named': 'c = d'}, build=False)
 
     assert len(actions) == 3
     assert str(actions[0]) == 'a = b'
@@ -181,7 +171,7 @@ def test_provides_access_to_constructed_objects_of_effect():
         model.b = spa.State()
         model.c = spa.State()
 
-        actions = spa.Actions('c = a * b')
+        actions = spa.Actions(('c = a * b',), build=False)
         bg, thalamus, constructed = actions.build()
 
         assert len(constructed[actions[0].effects[0]]) == 1